OCT torquetube, ReadResults routine, separation between modules in Row
Added a new OCT torquetube.

Added a readResults routine, that cleans for the materials of the module for 1up and 2up with torque tube. See JOURNAL for how to use it.

Added a varaible to modify the separation between panels. Named it "PSX" -- panel separation in X. This variable is introduced in the module creation, so the torque tube is longer than the module and there are no gaps in between torque tube, and also in the Scene itself so the modules are separated. PSX gets passed along many makeScene routines now.  Default 0.01

Signed-off-by: cdeline <chris.deline@nrel.gov>