Clean and deepclean routine. Gendaylit 2 FIX
Added another matcher to Clean routine - to detect inner borders. Not sure if it works on ALL cases must test more.

Just in case, left the "deepclean" routine, which can be automatic or request user input about the material.

Deepclean switches the sensors around, which Clean is not doing --> To do.

read1result turned back to previous without the pickle. Not sure if this was a merge issue or what but it's not working at the moment after the commit....  UGH :cold_sweat: