Gendaylit "Goodsky" Variable
Gendaylit still creates a sky for Sun below the horizon, but then Rtrace crashes. So added a variable to catch this and pass it forward.

It also checks if DNI is = or less to 0 becuase Gendaylit itself crashed on those instances for me.

Not sure if it's the most elegant way to handle these exceptions but it works :D