makeCustomObject and appendtoScene || And new Functionalities Journal Update
makeCustomObject and appendtoScene || And new Functionalities Journal Update

makeCustomObjectt so we can add foreing elements (like the block I use to know the 0,0 coordinate, or a cloth in just a specific area of the scene...)

appendtoScene so that the makeCustomObject can be easily added to the flow. (or removed by running makeScene again).

The Journal New Functionalities was updated, because before it was using "gendaylit2" which is now gendaylit. An example of how to add the cube with the new functions was added. PSX was also added to one of the examples that was not included before.