Opacity option for module
So, it doesn't give you opiton to set opacity but it creates a detailed module and THEN tells you the opacity for that. So if you increase spacing x and spacing y between the cells the opacity should decrease.  (I guess the term isn ot really opacity but Packaging Factor of the module)

I have some examples for the journal but I'll wait until its a bit more merged.