Implemented!
Implemented rotation around torque tube.
Just in case, it is set up as a boolean option. So you can set it to True and then the module and torque tube will shift up and to the side to keep the torque tube at the right distance from the ground and the unit element rotating around it.

Journal with some tests included for this. The sensors are also automatically updated in their location and match well with the materials/location of the panel.

Will run some more tests before Pull-request