scenex and sceney introduced, collectorwidth and later dependnecies on psx removed, orientation deprecated to a warning
scenex and sceney introduced, collectorwidth and later dependnecies on psx removed, orientation deprecated to a warning

the update to scenex made me realize we had to probably change this next line from x to scenex:
        text += '-i 1 -t %s %s 0 -rz %s ' %(-self.scenex*int(nMods/2), -pitch* (rowwanted - 1), 180-azimuth)

So I had to update the pytest itself (the translation would be 16.0 not 15.9 for 10 modules of 1.59 x +0.01 psx

Also I think because of this change or not sure what I'm getting 10 W/m2 extra on rear irradiance test_RadianceObj_1axis_gendaylit_end_to_end().... I need to check what's going on.

I still need to run some more sanity checks for this and for the 4 scenarios with the current changes but wanted to push them to beggin.