sensor module position updated for makescene (and improvements to this!!)
fix to scanning function so that it uses scenex not x for modwanted.
Added also a rotation to the sensors so it always scans the same module in the array (despite the array rotating...) so now for a 20 module array facing east, modwantd 1 will give mattype a0.0.a1.Pvmodule (and it is the southest module in teh row) and for when facing west, modwanted1 will give mattype a19.0.a1.Pvmodule (also the southest module in the row).  THIS IS IMPORTANT becuase otherwise we are querying a module at the north side of the array when facing West instead of a module at hte South side.

Just thought of a major breaktrough to do away with rotating the sensors too during cleaning :+1:  coming in a couple mins!