breakthrough in sensor positionnnnnnnnnnn
So, the geometry is stil messed up (it rotates 180, so panel A ends up in panel B when you are facing West for the trackers) so believing mattype and fronttype in the irr_results.csv is not an option, but,
Sensors are now measuring exaclty the same when facing the east and facing the west, so they can just be added nicely at the end of the routine for single axis tracking!!!!!!!!!!!!!!!

Also on deepclean we dont need to switch the panels and sensors anymore, just clean for undesired materials. I'll fix that next.

@cdeline  :D ... I 'll finish validating tho.