one more small update
I edited it on the online editor so I didn't see that it didn't like the spacing before the .format. So removed that. Added a parenthesis on a print statement that was missing.

I'm gonna try to re-base it to the new agnostic orientation branch ... somehow. Added the dictionary at least so it could work with my fixed and tracking landscape and portrait examples.