removing PSX ~ it was never in an official release
@cdeline  All looks good.
Just removed PSX - it was never in the official release, and it has had less than 3 weeks of existance so I think we are good with completely deprecating it.

Will check the tests next. It was not failing for me (except getting slightly higher irradiances in one... maybe tat's the one you ar ementioning. Will check!)