Case 1 parta fix
Case 1 :  Samping on Y
Sensors invert for Azimuths > 180 (so 180 to 225 degrees) properly on Y,
Discovered Zrear is not changing though!

Case 2: Sampling on X
Sesnors invert for Azimuths 225 to 360 properly
Zrear changes properly.  good job silvana

Gonna try to fix Z rear for Case 1, and test Case 3.

At some point, Case 1 and Case 2 should be MERGED, but calculating X increase and Y increase with the cosines and sine of Scene.y and the azimuth angle.