I'm a genius update
FIX TO #63

Sensors are working nicely. Still some more tests to run but,

they match, for ANY azimuth now. No more Case 1 and Case 2, they got simplified and all is good in the universe. It's kinda ingenious and I'm happy with it.

(no, not sure if modwanted works, we will see that too.) Still very exciting.

@cdeline