if pytest doesn't go to the mountain, hte mountain goes to pytest
Only single axis tracking is broken at the moment. but I am confident my hacks to pytest are necessary and justified and now it's passing most other tests.

Chankges to bifacial_radiance sensors fixed the Xstart that should have been 0. Updated all the other sensor positions because they are right. :D

btw at the moment, 'high tilt' is still scanning on x and y. It should work okay even for tilt = 90.  except the vector of scanning should shift. This is a good motive not to add another different x and y scan, but to modify the vector from 0 0 1 and 0 0 -1 to be the sin or cosine of the tilt. which ever it is supposed to be to be normal to the surface. @cdeline

Not sure hwat's going with single axis tilt. i haven't looked deep enough at it :-1: