backward compatibility makeScene1axis nMods nRows, and analysis1axis singleindex and sceneDict
makeScene1axis: Added an nMods and nRows input that gives a warning to the user, and checks to see if sceneDict has them - and if not assigns the value for them.

analysis1axis: switched the order between singelindex adn sceneDict for backward compatibility too.

moduleAnalysis: modified input value name from height to "hubheight", and added clearanceheight as an option to be passed. Also added specific values to all inputs jic because I think if you assign a specific value (like hubheight=None) all the variables afterwards also have to be defined. maybe. Anyways... this should be soon modified to be using sceneDict and moduleDict anyways!