hpc wait vs. none filetype 'fix'? AND modWanted rowWanted=0 Analysis fix, plus DOCUMENTATION
hpc check for None filetype. Not sure if this voids the hpc timecounter when it's waiting for the file?

Before when modwanted or rowanted = 0, a warning was given to the user but the value was not updated. Now value is updated to 1 for these cases - otherwise results are on the ground and sky :)  <-- right now the change is internal only., make sure to save this variable to the bigger scheme of analysisDict once that is implemented.

LOT OF DOCUMENTATION