gendaylit1axis startdate/enddate fix to trim trackerdictionary
trackerdictionary gets trimmed to entries that have a sky-file only ~ specially useful when doing startdate/enddate, since otherwise still the 4000 entries in the trackerdict stay there, with missing entries on sky.

Update to the content and text of the two main journals that explain how bifacial_radiance works. Also, cleanup to Mismatch journal for PVSC.