Model Chains in MAIN
Bifacial_radiacne.py: All changes here are to model chains. Ah, and it also reads a "config.py" file to get all the variables in dictionary form. Still under testing.

The goal is to get all the functions to take dictionaries, and not individual variables.

At this point, the process assumes that the config.py has all the right variables in all the right format. The Main section explores ALL of the possible options given currently by our software. It's kinda messy. Maybe this will help streamline stuff more.

plot_PVSC.py are some mismatch plots, still underworking.  It is just backup, so ignore.

config.ini <--- this is a good way to move forward after making config.py work. config.ini can be called from inside a function, and then we can do all the validation of the inputs, load the dictionaries inside of the code, and do the model chains. So this would be the neeeeeeeeeext stage.

Good job silvana. Maybe kinda.