fix issue #167 , add fix to #155
Fixe issue #167 by adding an actual check for the reindex -- current check only checked for trakcerdict format if reindex was not set.

Also renamed to load.exportTrackerDict load._exportTrackerDict to #155. Users should be calling demo.exportTrackerDict to export their tracker dictionaries, not the load._exportTrackerDict (which doesnt have the checks for reindex for example).