HPC-fixes
_tempTMY saves a file with the daydate now if available so it doesnt get overwritten/overreda.

importnat fix to sky generation on gendaylit1axis, because the for loop was going to endindex+1 before.

analysis1axis added input for daydate, so cumulative file adds for hte daydate. The other option is to turn off the cumulative file saving option for daydate? discuss

Main has the script to run the runJOB on multiprocessing. It's not worknig super good though. Summary:
1) 8 cores on my computer: It uses all cores on the computer
2) HPC: seems to be using both Nodes (for a 60 day simulatoin), but only starts printing stuff for one of the cores... the other one appears some of the results but later, does that means it's being assigned afterwards to the same initial node?