gendaylt Update for Albedo
Changes: 1) Made metdata input optional, as it is saved in self. @cdeline this is going to break some tests if the variables are not assigned (i.e. gendaylit(metdata = metdata, timeindex = timeindex), because the order of the variables changed to allow for metdata = None.

2) If metdata not passed, checks to see if it's stored in self.

3) Assigns the albedo from self.ground values .
Checks: If setGround() has not been run it warns the user and returns.
Checks: f self.ground values are different than metdata values and they are also not just 1 value (for a single albedo), it returns.