hpc module object name isseu
Not sure why but after yesterday's update, the module object is being saved in the main folder as object\\Bi60.rad, instead of on the objects folder, and even if I move it into the folder it is referneced wrong by teh scene.rad objects as objects\Bi60.rad and not objects/Bi60.rad    . Trying to specify the whole path like on the scene to see if this fixes that. Might need to roll back this change.