Mmonarch dev (#271)
* Add loading dewpoint and pressure from weather files, plus bug fix.

Store the dewpoint and pressure in MetObj. This data is used when
generating dni/dhi spectrum using SEDES2. Currently, this has only
been tested with loading epw files. Also fix bug related to
running simulations with differing RGB channels.

* Create a simple interface for using spectral data.

Added a class to act as an interface when dealing with values that
vary based on wavelength, such as albedo or solar irradiance.
Included two utilities to load spectral data for albedo from smarts
and spectral data for dhi/dni from SEDES2.

* Add method to get irradiance spectra from SMARTS

When retrieving a list of values from a spectral
property, a numpy array is returned instead of a list

* update JOSS, Zenodo on README.md

Co-authored-by: Mark Monarch <Mark.Monarch@nrel.gov>
Co-authored-by: cdeline <chris.deline@nrel.gov>