_saveTempTMY and coerce_year
If coerce_year is not passed, the function will now check if the weather file needs to have the year coerced anyway (such as TMYs with non-sequential datetimes) or apply the datetime index as it is read (if the file spans multiple years). This prevents datetime errors cropping up when the argument is not passed or from excluding data with the indexmask should the file finish with a year chronologically after the initial datetime index. Default coerce_year is 2001.