unfinished step on the right direction
TMY and EPW routines have been turned internal and reduced to just reading the files (and correctig) for PVlib, so now all the year coersion, masking (reducing times) and indexing issues with field data and etc are handled on the above function and are the same for both.

Also, internal epwfile that pointed to the temporal file to be used by gencumsky modified in name throughout.