removed readTrackerAngles() and pytest
Decided not to do readTrackerAngles because metObj saves a reduced metdata without the night hours and it's going to be hard to align the data or would have to rerun metdata file, so let's just make users add their trackerdata to the weatherfile. If this function ends up being needed it can be added once we have a real use example/need.

also added pytest