COOL STUFF
load._exportTrackerDict - removed the bifacial ratio calculation because that's not where to do that. Added a bunch of other columns.  Might need to column proof if the trackerdict is attempted to be exported BEFORE running the performance calculations

main
-Changed calculatePerformanceModule to calculateResults, as now it includes electrical mismatch and clean front and rear irradiances and the bifaciality factor. all awesome on this new function.

mismatch mad function: corrected to drop nans

performance
updates to calculatePerformanceModule so there is no interpolation that skews the results of the mean of irradiances,
standarized names throghout,
calculating now BGG and BGE