Generating Spectra
New method `RadianceObject.generate_spectra`. Generates spectra directly from radiance object and saves the generated files in the "spectra" folder in the radiance scene directory. When passing `scale_albedo_nonspectral_sim` the `radianceObject.metdata.albedo` is automatically updated with the new weighted albedo values so there is no need to re-read weather to continue the simulation. Additionally, a datetime-indexed csv with the weighted albedo is placed in the "spectra" folder.
Spectra generation will now process for the entire length of metdata within the radiance object, so whatever datetime-range is specified in `readweather` will be processed by `generate_spectra`.