Tutorial 2 and load.py
updated tutorial 2 to match the changes in analysis1axis and performance.calculateResults. Updated load.read1Result to detect if the file read already contains an index. If so, it applies that index to the dataframe. This prevents the appearance of columns like "Unnamed: 0" from showing up in your result dataframe.